Move computation to base class
